feat(tools)!: grounding -> tako_answer on GA /api/v1/answer
BREAKING: tool renamed; endpoint migrated from /api/v1/grounding to
/api/v1/answer; output is {answer,cards,web_results,request_id} (drops
tako_selected/confidence).
