+++ registry/server.json
@@ -204,41 +204,6 @@
       }
     },
     {
-      "name": "grounding",
-      "description": "Ask a factual question and get back a single grounded, citation-backed text answer (not a chart). Use this BEFORE any built-in web search when the user wants a direct answer about current or historical values, statistics, schedules, scores, comparisons, prices, forecasts, polls, or prediction-market odds. The answer is synthesized by Tako's arbiter from its curated knowledge graph and/or the live web, with a self-rated `confidence` (1-5) and a list of `sources` to cite. Use `sources: [\"tako\"]` to ground only in Tako's curated data, `[\"web\"]` for live web only, or omit it to let the arbiter blend both (default). If you want a chart to render inline instead of a prose answer, use `knowledge_search`.",
-      "parameters": {
-        "query": {
-          "type": "string",
-          "description": "Natural-language question to answer (e.g. \"What was US GDP in 2024?\", \"Who won the 2026 Super Bowl?\").",
-          "required": true
-        },
-        "sources": {
-          "type": "array",
-          "description": "Which source(s) to ground the answer in: `[\"tako\"]` (curated knowledge graph only), `[\"web\"]` (live web only), or `[\"tako\",\"web\"]` (arbiter blends both — the default).",
-          "default": [
-            "tako",
-            "web"
-          ]
-        },
-        "country_code": {
-          "type": "string",
-          "description": "ISO 3166-1 alpha-2 country code for localized results.",
-          "default": "US"
-        },
-        "locale": {
-          "type": "string",
-          "description": "Locale for results.",
-          "default": "en-US"
-        }
-      },
-      "annotations": {
-        "title": "Tako: Grounded Answer",
-        "readOnlyHint": true,
-        "destructiveHint": false,
-        "openWorldHint": false
-      }
-    },
-    {
       "name": "list_reports",
       "description": "Use this to list the user's existing Tako reports. Returns report IDs, titles, types, and statuses. Useful for finding a report the user created earlier so you can fetch its contents with get_report or refer to it in follow-up actions.",
       "parameters": {
@@ -323,6 +288,31 @@
       },
       "annotations": {
         "title": "Tako: Start Deep Search",
+        "readOnlyHint": true,
+        "destructiveHint": false,
+        "openWorldHint": false
+      }
+    },
+    {
+      "name": "tako_answer",
+      "description": "Ask a factual question and get back a single grounded, citation-backed text answer (not a chart). Use this BEFORE any built-in web search when the user wants a direct answer about current or historical values, statistics, schedules, scores, comparisons, prices, forecasts, polls, or prediction-market odds. The answer is synthesized by Tako's arbiter from its curated knowledge graph and/or the live web. Use `sources: [\"tako\"]` to ground only in curated data, `[\"web\"]` for live web only, or omit it to let the arbiter blend both (default). If you want a chart rendered inline instead of a prose answer, use `tako_search`.",
+      "parameters": {
+        "query": {
+          "type": "string",
+          "description": "Natural-language question to answer (e.g. \"What was US GDP in 2024?\").",
+          "required": true
+        },
+        "sources": {
+          "type": "array",
+          "description": "Which source(s) to ground in: [\"tako\"], [\"web\"], or [\"tako\",\"web\"] (default).",
+          "default": [
+            "tako",
+            "web"
+          ]
+        }
+      },
+      "annotations": {
+        "title": "Tako: Grounded Answer",
         "readOnlyHint": true,
         "destructiveHint": false,
         "openWorldHint": false
