+++ workers/src/tools/tako_answer.test.ts
@@ -1,0 +1,170 @@
+/**
+ * Tests for the `tako_answer` tool.
+ *
+ * The handler is a single POST + re-parse: it maps the tool input onto
+ * the backend's `/api/v1/answer/` request body (v3 SearchRequest shape:
+ * top-level `query` + `source_indexes`, NOT `inputs.text`) and re-validates
+ * the response through the output schema. The interesting behavior is:
+ *   - request mapping (`query`→`query`, `sources`→`source_indexes`)
+ *   - the defensive defaulting of missing fields (cards/web_results → [])
+ *   - the loud failure on a mis-shaped backend payload
+ *   - the absence of grounding-era fields (tako_selected, confidence)
+ */
+import { afterEach, describe, expect, it, vi } from "vitest";
+
+import type { Env } from "../env.js";
+import type { ToolContext } from "./types.js";
+import takoAnswer from "./tako_answer.js";
+import {
+  bodyOf,
+  jsonResponse,
+  mockFetchSequence,
+  noopSendProgress,
+  requestFrom,
+} from "./__test_helpers.js";
+
+const ENV: Env = { DJANGO_BASE_URL: "https://staging.trytako.com" };
+const CTX: ToolContext = {
+  token: "sk-test",
+  env: ENV,
+  sendProgress: noopSendProgress,
+  client: "claude",
+};
+
+const FULL_RESPONSE = {
+  answer: "US GDP was about $29 trillion in 2024.",
+  cards: [
+    {
+      card_id: "abc123",
+      title: "US GDP",
+      description: "Gross Domestic Product of the United States",
+      webpage_url: "https://trytako.com/charts/us-gdp",
+      image_url: "https://trytako.com/api/v1/image/abc123/",
+      embed_url: "https://trytako.com/embed/abc123/",
+    },
+  ],
+  web_results: [
+    { title: "US GDP 2024", url: "https://example.com/gdp", snippet: "..." },
+  ],
+  request_id: "req-1",
+};
+
+afterEach(() => {
+  vi.unstubAllGlobals();
+  vi.restoreAllMocks();
+});
+
+describe("tako_answer handler", () => {
+  it("tool name is tako_answer", () => {
+    expect(takoAnswer.name).toBe("tako_answer");
+  });
+
+  it("maps query→query and sources→source_indexes, hits /api/v1/answer/", async () => {
+    const fetchMock = mockFetchSequence([jsonResponse(200, FULL_RESPONSE)]);
+
+    const out = await takoAnswer.handler(
+      {
+        query: "What was US GDP in 2024?",
+        sources: ["tako", "web"],
+      },
+      CTX,
+    );
+
+    const req = requestFrom(fetchMock.mock.calls[0]);
+    expect(req.url).toBe("https://staging.trytako.com/api/v1/answer/");
+    const body = await bodyOf(req);
+    // v3 SearchRequest: top-level `query` (NOT inputs.text)
+    expect(body.query).toBe("What was US GDP in 2024?");
+    expect(body.source_indexes).toEqual(["tako", "web"]);
+    // grounding-era nested inputs must NOT be present
+    expect(body.inputs).toBeUndefined();
+
+    expect(out.answer).toContain("$29 trillion");
+    expect(out.cards).toHaveLength(1);
+    expect(out.cards[0]?.card_id).toBe("abc123");
+    expect(out.web_results).toHaveLength(1);
+    expect(out.request_id).toBe("req-1");
+  });
+
+  it("defaults missing optional fields to empty arrays rather than leaking undefined", async () => {
+    // A minimal valid backend payload — only the required scalar. The
+    // handler should default cards/web_results to [].
+    mockFetchSequence([
+      jsonResponse(200, {
+        answer: "No strong match.",
+        request_id: "req-2",
+      }),
+    ]);
+
+    const out = await takoAnswer.handler(
+      { query: "obscure query", sources: ["tako"] },
+      CTX,
+    );
+
+    expect(out.cards).toEqual([]);
+    expect(out.web_results).toEqual([]);
+    expect(out.answer).toBe("No strong match.");
+    expect(out.request_id).toBe("req-2");
+  });
+
+  it("output does NOT contain grounding-era fields (tako_selected, confidence)", async () => {
+    mockFetchSequence([jsonResponse(200, FULL_RESPONSE)]);
+
+    const out = await takoAnswer.handler(
+      { query: "test", sources: ["tako"] },
+      CTX,
+    ) as Record<string, unknown>;
+
+    expect(out.tako_selected).toBeUndefined();
+    expect(out.confidence).toBeUndefined();
+  });
+
+  it("throws an actionable error when the backend returns an unexpected shape (cards not an array)", async () => {
+    mockFetchSequence([
+      jsonResponse(200, {
+        answer: "ok",
+        cards: "not-an-array",
+        web_results: [],
+        request_id: "req-bad",
+      }),
+    ]);
+
+    await expect(
+      takoAnswer.handler({ query: "q", sources: ["tako", "web"] }, CTX),
+    ).rejects.toThrow(/unexpected shape/);
+  });
+});
+
+describe("tako_answer input schema", () => {
+  it("defaults sources to both tako and web", () => {
+    const parsed = takoAnswer.inputSchema.parse({ query: "hello" });
+    expect(parsed.sources).toEqual(["tako", "web"]);
+  });
+
+  it("rejects an empty sources array", () => {
+    expect(() =>
+      takoAnswer.inputSchema.parse({ query: "hello", sources: [] }),
+    ).toThrow();
+  });
+
+  it("rejects an unknown source", () => {
+    expect(() =>
+      takoAnswer.inputSchema.parse({ query: "hello", sources: ["bing"] }),
+    ).toThrow();
+  });
+
+  it("does NOT have country_code or locale fields (dropped from v3 shape)", () => {
+    // These are extra fields — zod strips them by default; parse still succeeds
+    // but the output should not contain them as required fields
+    const parsed = takoAnswer.inputSchema.safeParse({
+      query: "hello",
+      country_code: "US",
+      locale: "en-US",
+    });
+    expect(parsed.success).toBe(true);
+    if (parsed.success) {
+      expect((parsed.data as Record<string, unknown>).country_code).toBeUndefined();
+      expect((parsed.data as Record<string, unknown>).locale).toBeUndefined();
+    }
+  });
+});
